Crate spin_factors

Re-exports

pub use crate::runtime_config::FactorRuntimeConfigSource;

pub use crate::runtime_config::RuntimeConfigSourceFinalizer;

pub use anyhow;

pub use serde;

pub use wasmtime;

Modules

runtime_config

Structs

App

An App holds loaded configuration for a Spin application.

AppComponent

An AppComponent holds configuration for a Spin application component.

ConfigureAppContext

FactorData

Helper type to satisfy the D: HasData implementations needed by wasmtime’s generated add_to_linker functions.

PrepareContext

A PrepareContext is passed to Factor::prepare.

Enums

Error

Traits

AsInstanceState

Factor

A contained (i.e., “factored”) piece of runtime functionality.

FactorInstanceBuilder

A builder for a Factor’s per instance state.

HasInstanceBuilder

Allows querying an InstanceBuilders for a particular Factor’s InstanceBuilder.

InitContext

An InitContext is passed to Factor::init, giving access to the global common wasmtime::component::Linker.

RuntimeFactors

A collection of Factors that are initialized and configured together.

RuntimeFactorsInstanceState

Get the state of a particular Factor from the overall InstanceState

SelfInstanceBuilder

A helper trait for when the type implementing FactorInstanceBuilder is also the instance state.

Type Aliases

FactorInstanceState

The instance state of the given Factor F.

Result

Result wrapper type defaulting to use Error.

Derive Macros

RuntimeFactors