spin_factors

Trait FactorInstanceBuilder

pub trait FactorInstanceBuilder: Any {
    type InstanceState: Send + 'static;

    // Required method
    fn build(self) -> Result<Self::InstanceState>;
}

A builder for a Factor’s per instance state.

Required Associated Types

type InstanceState: Send + 'static

The per instance state of the factor.

This is equivalent to the existing HostComponent::Data and ends up being stored in the wasmtime::Store. Any bindgen traits for this factor will be implemented on this type.

Required Methods

fn build(self) -> Result<Self::InstanceState>

Build the per instance state of the factor.